# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "이화여대 창업 교육 LMS API",
        "title": "EarnLearning LMS API",
        "contact": {},
        "version": "1.0"
    },
    "host": "earnlearning.com",
    "basePath": "/api",
    "paths": {
...
        "/admin/companies": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "관리자용: 모든 회사 목록 조회",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "전체 회사 목록",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/companies": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "새 회사를 설립 (자본금 필요)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Company"
                ],
                "summary": "회사 설립",
                "parameters": [
                    {
                        "description": "회사 설립 정보",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.CreateCompanyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.APIResponse"
                        }
                    }
                }
            }
        },
        "/companies/mine": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "내가 소유한 회사 목록 조회",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Company"
                ],
                "summary": "내 회사 목록",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.APIResponse"
                        }
                    }
                }
            }
        },
        "/companies/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "특정 회사의 상세 정보 조회",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Company"
                ],
                "summary": "회사 상세 조회",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "회사 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.APIResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "회사 설명, 로고 등 수정 (소유자만)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Company"
                ],
                "summary": "회사 정보 수정",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "회사 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "수정 정보",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.UpdateCompanyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.APIResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.APIResponse"
                        }
                    }
                }
            }
        },
        "/companies/{id}/business-card": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "회사 명함 조회",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Company"
                ],
                "summary": "명함 조회",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "회사 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "회사 명함 생성 또는 수정",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Company"
                ],
                "summary": "명함 생성/수정",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "회사 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "명함 정보",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.BusinessCardRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.APIResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_interfaces_http_handler.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "INVALID_INPUT"
                },
                "message": {
                    "type": "string",
                    "example": "잘못된 입력입니다"
                }
            }
        },
        "internal_interfaces_http_handler.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/internal_interfaces_http_handler.APIError"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "internal_interfaces_http_handler.BusinessCardRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "ceo@company.com"
                },
                "name": {
                    "type": "string",
                    "example": "홍길동"
                },
                "phone": {
                    "type": "string",
                    "example": "010-1234-5678"
                },
                "template": {
                    "type": "string",
                    "example": "modern"
                },
                "title": {
                    "type": "string",
                    "example": "CEO"
                }
            }
        },
...
        "internal_interfaces_http_handler.CreateCompanyRequest": {
            "type": "object",
            "properties": {
                "capital": {
                    "type": "integer",
                    "example": 100000
                },
                "description": {
                    "type": "string",
                    "example": "AI 기반 서비스"
                },
                "name": {
                    "type": "string",
                    "example": "테크스타트업"
                }
            }
        },
...
        "internal_interfaces_http_handler.UpdateCompanyRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "수정된 설명"
                },
                "logo_url": {
                    "type": "string",
                    "example": "/uploads/logo.png"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}